// src/modules/inventory/reports/finish-store/date-wise-fabric-purchase-receive-register-report/date-wise-fabric-purchase-receive-register-report-type.ts
export type DateWiseFabricPurchaseReceiveRegisterReportType = {
  RCV_CHALLAN_DATE: string;
  BUYER: string;
  WORK_ORDER_NO: string;
  SUPPLIER: string;
  FABRIC: string;
  LC_NO: string;
  RCV_CHALLAN: string;
  RATE: number;
  UOM: string;
  CURRENCYCODE: string;
  QUANTITY: number;
  ROLL_QTY: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
};